# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Interview Platform API",
    "version": "1.0.0",
    "description": "API for AI-powered interview platform. Base URL: `/api/v1`.\n\nAuthentication: Bearer JWT (access_token). Refresh token via HttpOnly Cookie.\n\nStandard response envelope:\n- Success: `{ success: true, data, meta, request_id }`\n- Error: `{ success: false, error: { code, message, details }, request_id }`",
    "contact": {
      "name": "Khôi - Backend Core",
      "email": "khoi@example.com"
    }
  },
  "servers": [
    { "url": "/api/v1", "description": "API v1" }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new account",
        "description": "Creates a recruiter account. Returns access_token and sets refresh_token as HttpOnly Cookie.",
        "operationId": "authRegister",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "full_name"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "recruiter@example.com" },
                  "password": { "type": "string", "minLength": 8, "example": "Password123!" },
                  "full_name": { "type": "string", "example": "Nguyễn Văn A" },
                  "role": { "type": "string", "enum": ["recruiter", "candidate"], "default": "recruiter" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AuthResponse" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AuthResponse" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user profile",
        "security": [{ "BearerAuth": [] }],
        "operationId": "authMe",
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/UserMe" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "description": "Reads refresh_token from HttpOnly Cookie. Returns new access_token and rotates refresh_token via Set-Cookie.",
        "operationId": "authRefresh",
        "responses": {
          "200": {
            "description": "New tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout (revoke current refresh token family)",
        "security": [{ "BearerAuth": [] }],
        "operationId": "authLogout",
        "responses": {
          "200": { "description": "Logged out" }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout all sessions",
        "security": [{ "BearerAuth": [] }],
        "operationId": "authLogoutAll",
        "responses": {
          "200": {
            "description": "All sessions revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" },
                        "revoked_sessions": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "UserMe": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "full_name": { "type": "string" },
          "role": { "type": "string" },
          "avatar_url": { "type": "string", "nullable": true },
          "companies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" },
                "role": { "type": "string" }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "AuthResponse": {
        "description": "Auth tokens + user info",
        "headers": {
          "Set-Cookie": {
            "description": "HttpOnly refresh_token cookie",
            "schema": { "type": "string" }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean" },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" },
                        "full_name": { "type": "string" },
                        "role": { "type": "string" },
                        "status": { "type": "string" }
                      }
                    },
                    "access_token": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Invalid or missing credentials",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "enum": [false] },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "enum": ["UNAUTHORIZED"] },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Resource conflict (duplicate)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "enum": [false] },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "enum": ["CONFLICT"] },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}